n6/requirements.txt
asgiref==3.6.0
astroid==2.12.13
autopep8==2.0.1
colorama==0.4.6
dill==0.3.6
Django==4.1.4
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
isort==5.11.4
lazy-object-proxy==1.8.0
Markdown==3.4.1
mccabe==0.7.0
platformdirs==2.6.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pylint==2.15.9
pytz==2022.7
sqlparse==0.4.3
tomlkit==0.11.6
tzdata==2022.7
wrapt==1.14.1
